Complete class Reservation

+++ src/level2/HotelReservationSystem/Room.java
@@ -25,7 +25,7 @@ public class Room {
     }
 
 
-   public double getPrice(int N){
+   public double getPrice(long N){
         return N*pricePerNight;
    }
